hsh
Module algorithms

The algorithms module contains the password hashing algorithms.

Modules

argon2i

The argon2i module contains the Argon2i password hashing algorithm.

bcrypt

The bcrypt module contains the Bcrypt password hashing algorithm.

scrypt

The scrypt module contains the Scrypt password hashing algorithm.